+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to undo a rename/relink instantly after it is performed.
 - `title-tidy undo` to undo rename/relink after the title tidy session has closed.
 - viewport support to stat panels to fix TUI distortion when terminal is too short.
+- Allow configuring logging parameters via the config UI.
+### Updated
+- Config demo gif to show new section. 
 
 ## [v1.5.1] - 2025-08-26
 ### Fixed